feat: updated rnd
